Refactor imports to use the new relationship_manager module and remove unnecessary references to RelationshipManager in model files.
