+++ tools/yaml2x/libs/freee_a11y_gl/freee_a11y_gl/models/__init__.py
@@ -1,14 +1,14 @@
 """Models package for a11y-guidelines.
 
 This package contains models for:
-- Base functionality (BaseModel, RelationshipManager)
+- Base functionality (BaseModel)
 - Content (Category, Guideline)
 - Checks (Check, CheckTool, etc.)
 - FAQs (Faq, FaqTag)
 - References (WcagSc, InfoRef)
 - axe-core integration (AxeRule)
 """
-from .base import BaseModel, RelationshipManager
+from .base import BaseModel
 from .content import Category, Guideline
 from .check import Check, CheckTool, Condition, Procedure, Implementation, Method
 from .faq import Faq, FaqTag
@@ -18,7 +18,6 @@ from .axe import AxeRule
 __all__ = [
     # Base
     'BaseModel',
-    'RelationshipManager',
     
     # Content
     'Category',
