+++ tools/yaml2x/libs/freee_a11y_gl/freee_a11y_gl/models/content.py
@@ -1,7 +1,8 @@
 """Content models for a11y-guidelines."""
 from typing import Dict, List, Optional, Any
 from dataclasses import dataclass
-from .base import BaseModel, RelationshipManager
+from .base import BaseModel
+from ..relationship_manager import RelationshipManager
 from ..config import Config, LanguageCode
 from ..utils import uniq
